cleanup(16): - remove unused hello world code refs: #16
